+++ src/components/main/Squeak.js
@@ -1,36 +1,28 @@
 import { createClient } from '@supabase/supabase-js'
-import React, { useEffect, useRef, useState } from 'react'
+import React, { useRef } from 'react'
 import root from 'react-shadow/styled-components'
 import { Provider as SupabaseProvider } from 'react-supabase'
 import { Provider as OrgProvider } from '../../context/org'
-import getBackgroundColor from '../../util/getBackgroundColor'
-import lightOrDark from '../../util/lightOrDark'
+import { Provider as UserProvider } from '../../context/user'
 import Questions from '../Questions'
 import { Theme } from '../Theme'
 
 export const Squeak = ({ apiKey, url, apiHost, organizationId }) => {
-  const [darkMode, setDarkMode] = useState(null)
+  const supabase = createClient(url, apiKey)
   const containerRef = useRef()
 
-  const supabase = createClient(url, apiKey)
-
-  useEffect(() => {
-    if (containerRef.current) {
-      const color = getBackgroundColor(containerRef.current)
-      setDarkMode(lightOrDark(color) === 'dark')
-    }
-  }, [])
-
   return (
-    <SupabaseProvider value={supabase}>
-      <OrgProvider value={{ organizationId, apiHost }}>
-        <root.div ref={containerRef}>
-          <Theme dark={darkMode} />
-          <div className='squeak'>
-            <Questions />
-          </div>
-        </root.div>
-      </OrgProvider>
-    </SupabaseProvider>
+    <root.div ref={containerRef}>
+      <SupabaseProvider value={supabase}>
+        <OrgProvider value={{ organizationId, apiHost }}>
+          <UserProvider>
+            <Theme containerRef={containerRef} />
+            <div className='squeak'>
+              <Questions />
+            </div>
+          </UserProvider>
+        </OrgProvider>
+      </SupabaseProvider>
+    </root.div>
   )
 }
